test: add a todo into localStorage
Tests were conducted in browser and not using Jest testing.
Jest testing does not seem to simulate Web Storage API.
A browser-run script tested adding a todo object from createToDo
factory, store it in localStorage, retrieve it from localStorage,
and lastly display the retrieved object in console.log.
The retrieved object was exactly the same as the initial object.