refactor: create global event handler module
Created a dataHandler.js module that holds functions that add event
listeners to the DOM. This is necessary because the event listeners for
the project and task forms require access to globally shared data.
These global variables, such as projectsArray, will be stored and
manipulated within the dataHandler.js module.